projects/agentic-ai/ai-agents-project/requirements.txt
langchain>=1.2.0
langchain-google-genai>=4.2.0
langchain-tavily>=0.2.0
langgraph>=1.0.0
python-dotenv>=1.0.0
requests>=2.32.0
beautifulsoup4>=4.12.0
google-genai>=1.0.0
pandas>=2.0.0
streamlit>=1.40.0
pyyaml>=6.0.0
pytest>=9.0.0